refactoring organization and application names
