+++ cva-storage/cva-storage-mongodb/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.gel.cva</groupId>
+        <artifactId>cva-storage</artifactId>
+        <version>1.0</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <artifactId>cva-storage-mongodb</artifactId>
+    <version>1.0</version>
+    <packaging>jar</packaging>
+
+    <dependencies>
+        <!-- Intra-module dependencies -->
+        <dependency>
+            <groupId>org.gel.cva</groupId>
+            <artifactId>cva-storage-core</artifactId>
+            <version>1.0</version>
+        </dependency>
+    </dependencies>
+
+</project>
