--- ntbcvariantbox-storage/pom.xml
@@ -1,22 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.tobedefined.ntbcvariantbox</groupId>
-        <artifactId>ntbcvariantbox</artifactId>
-        <version>1.0</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <!-- Version must be set to a literal, variables cannot be used in nested modules -->
-    <artifactId>ntbcvariantbox-storage</artifactId>
-    <version>1.0</version>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>ntbcvariantbox-storage-core</module>
-        <module>ntbcvariantbox-storage-mongodb</module>
-    </modules>
-
-</project>
