+++ pom.xml
@@ -4,13 +4,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.tobedefined.ntbcvariantbox</groupId>
-    <artifactId>ntbcvariantbox</artifactId>
+    <groupId>org.gel.cva</groupId>
+    <artifactId>cva</artifactId>
     <version>1.0</version>
     <packaging>pom</packaging>
 
     <modules>
-        <module>ntbcvariantbox-storage</module>
+        <module>cva-storage</module>
     </modules>
 
     <properties>
